Moved e2e test utils to its files
