+++ test/e2e/ddfs_consts.py
@@ -1,0 +1,8 @@
+# Keep in sync with module/consts.h
+class DdfsConsts:
+    CLUSTER_UNUSED = 0
+    CLUSTER_NOT_ASSIGNED = 0xfffffffe
+    CLUSTER_EOF = 0xffffffff
+
+    FILE_ATTR = 1
+    DIR_ATTR = 2
